Add UioDevice::get_mapping_info()
This is a replacement for get_map_info(), which is slightly broken.

Bugs (in the old version):

The Vec is not guaranteed to be ordered, because fs::read_dir doesn't
guarantee an ordering. This makes the API easy to misuse, because
naïvely using `.iter().enumerate()` will produce an index which might
not match the string name – e.g. `(2, "map0")`.

The "mapping name" passed in the string is not actually the mapping name
as described by `.../maps/map*/name`, and there was no way to get it.

I've marked the old method as deprecated, with a note to use this new
one, so there is no breaking change.